feat: style movies layout with new MoviesContainer component
